split side quest path components
